requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
SQLAlchemy==2.0.31
python-multipart==0.0.9
pydantic==2.8.2
passlib[bcrypt]==1.7.4
python-jose==3.3.0
python-docx==1.1.2
pdfminer.six==20240706
python-dotenv==1.0.1
starlette==0.37.2
httpx==0.27.2

